intrinsic_dimension/intrinsic_dim/requirements.txt
numpy
h5py
IPython
colorama
setproctitle
tensorflow-gpu
keras
orderedset
scikit-learn